[Task #1652] Remove term assignments-related code TermOccurrenceService.
Merge the TermOccurrenceService interface with implementation. Remove tags code from resource management. Remove assignment generation from AnnotationGenerator.
