+++ src/main/java/cz/cvut/kbss/termit/service/business/ResourceService.java
@@ -20,7 +20,6 @@ import cz.cvut.kbss.termit.event.FileRenameEvent;
 import cz.cvut.kbss.termit.exception.InvalidParameterException;
 import cz.cvut.kbss.termit.exception.NotFoundException;
 import cz.cvut.kbss.termit.exception.UnsupportedAssetOperationException;
-import cz.cvut.kbss.termit.model.Term;
 import cz.cvut.kbss.termit.model.TextAnalysisRecord;
 import cz.cvut.kbss.termit.model.Vocabulary;
 import cz.cvut.kbss.termit.model.changetracking.AbstractChangeRecord;
@@ -80,17 +79,6 @@ public class ResourceService
     }
 
     /**
-     * Sets terms with which the specified target resource is annotated.
-     *
-     * @param target   Target resource
-     * @param termUris Identifiers of terms annotating the resource
-     */
-    @Transactional
-    public void setTags(Resource target, Collection<URI> termUris) {
-        repositoryService.setTags(target, termUris);
-    }
-
-    /**
      * Removes resource with the specified identifier.
      * <p>
      * Resource removal also involves cleanup of annotations and term occurrences associated with it.
@@ -105,16 +93,6 @@ public class ResourceService
         final Resource actualToRemove = getRequiredReference(toRemove.getUri());
         documentManager.remove(actualToRemove);
         repositoryService.remove(actualToRemove);
-    }
-
-    /**
-     * Gets terms the specified resource is annotated with.
-     *
-     * @param resource Annotated resource
-     * @return List of terms annotating the specified resource
-     */
-    public List<Term> findTags(Resource resource) {
-        return repositoryService.findTags(resource);
     }
 
     /**
@@ -215,7 +193,7 @@ public class ResourceService
         } else {
             persist(file);
         }
-        if ( !getReference(document.getUri()).isPresent() ) {
+        if (!getReference(document.getUri()).isPresent()) {
             persist(document);
         } else {
             update(doc);
@@ -225,7 +203,7 @@ public class ResourceService
     /**
      * Removes the file. The file is detached from the document and removed, together with its content.
      *
-     * @param file     The file to add and save
+     * @param file The file to add and save
      * @throws UnsupportedAssetOperationException If the specified resource is not a Document
      */
     @Transactional
@@ -236,7 +214,7 @@ public class ResourceService
             throw new InvalidParameterException("File was not attached to a document.");
         } else {
             doc.removeFile(file);
-            if ( repositoryService.getReference(doc.getUri()).isPresent() ) {
+            if (repositoryService.getReference(doc.getUri()).isPresent()) {
                 update(doc);
             }
         }
@@ -349,13 +327,13 @@ public class ResourceService
             final Resource original = findRequired(instance.getUri());
             if (!Objects.equals(original.getLabel(), instance.getLabel())) {
                 return Optional.of(new FileRenameEvent((File) instance, original.getLabel(),
-                    instance.getLabel()));
+                        instance.getLabel()));
             }
-        } else if (instance instanceof Document ) {
+        } else if (instance instanceof Document) {
             final Resource original = findRequired(instance.getUri());
             if (!Objects.equals(original.getLabel(), instance.getLabel())) {
                 return Optional.of(new DocumentRenameEvent((Document) instance, original.getLabel(),
-                    instance.getLabel()));
+                        instance.getLabel()));
             }
         }
         return Optional.empty();
