+++ src/main/java/cz/cvut/kbss/termit/service/document/AnnotationGenerator.java
@@ -21,13 +21,10 @@ import cz.cvut.kbss.termit.exception.AnnotationGenerationException;
 import cz.cvut.kbss.termit.model.Asset;
 import cz.cvut.kbss.termit.model.Term;
 import cz.cvut.kbss.termit.model.assignment.OccurrenceTarget;
-import cz.cvut.kbss.termit.model.assignment.TermAssignment;
 import cz.cvut.kbss.termit.model.assignment.TermOccurrence;
 import cz.cvut.kbss.termit.model.resource.File;
 import cz.cvut.kbss.termit.model.selector.Selector;
 import cz.cvut.kbss.termit.persistence.dao.TermOccurrenceDao;
-import cz.cvut.kbss.termit.service.repository.TermAssignmentRepositoryService;
-import cz.cvut.kbss.termit.util.Configuration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -35,10 +32,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.io.InputStream;
-import java.net.URI;
 import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 /**
  * Creates annotations (term occurrences) for vocabulary terms.
@@ -51,26 +46,18 @@ public class AnnotationGenerator {
 
     private static final Logger LOG = LoggerFactory.getLogger(AnnotationGenerator.class);
 
-    private final Configuration configuration;
-
     private final TermOccurrenceDao termOccurrenceDao;
 
     private final DocumentManager documentManager;
 
-    private final TermAssignmentRepositoryService assignmentService;
-
     private final TermOccurrenceResolvers resolvers;
 
     @Autowired
-    public AnnotationGenerator(Configuration configuration,
-                               TermOccurrenceDao termOccurrenceDao,
+    public AnnotationGenerator(TermOccurrenceDao termOccurrenceDao,
                                DocumentManager documentManager,
-                               TermAssignmentRepositoryService assignmentService,
                                TermOccurrenceResolvers resolvers) {
-        this.configuration = configuration;
         this.termOccurrenceDao = termOccurrenceDao;
         this.documentManager = documentManager;
-        this.assignmentService = assignmentService;
         this.resolvers = resolvers;
     }
 
@@ -87,7 +74,6 @@ public class AnnotationGenerator {
         occurrenceResolver.parseContent(content, source);
         final List<TermOccurrence> occurrences = occurrenceResolver.findTermOccurrences();
         saveOccurrences(occurrences, source);
-        generateAssignments(occurrences, source);
         saveAnnotatedContent(source, occurrenceResolver.getContent());
         LOG.trace("Finished generating annotations for file {}.", source);
     }
@@ -107,9 +93,9 @@ public class AnnotationGenerator {
         final List<TermOccurrence> existing = termOccurrenceDao.findAllTargeting(source);
         occurrences.stream().filter(o -> isNew(o, existing))
                    .filter(o -> !o.getTerm().equals(source.getUri())).forEach(o -> {
-            o.addType(cz.cvut.kbss.termit.util.Vocabulary.s_c_navrzeny_vyskyt_termu);
-            termOccurrenceDao.persist(o);
-        });
+                       o.addType(cz.cvut.kbss.termit.util.Vocabulary.s_c_navrzeny_vyskyt_termu);
+                       termOccurrenceDao.persist(o);
+                   });
     }
 
     /**
@@ -141,15 +127,6 @@ public class AnnotationGenerator {
             }
         }
         return true;
-    }
-
-    private void generateAssignments(List<TermOccurrence> occurrences, File source) {
-        LOG.trace("Creating term assignments for file {}.", source);
-        final double minScore = Double.parseDouble(configuration.getTextAnalysis().getTermAssignmentMinScore());
-        final Set<URI> termsToAssign = occurrences.stream()
-                                                  .filter(o -> o.getScore() != null && o.getScore() >= minScore)
-                                                  .map(TermAssignment::getTerm).collect(Collectors.toSet());
-        assignmentService.addToResourceSuggested(source, termsToAssign);
     }
 
     private void saveAnnotatedContent(File file, InputStream input) {
